backend, frontend, feature: htmx bootstrap modal to display topic details.
 - It was applied to the report, tracker, topics, and index pages.

The HTML code on the index page is a bit tricky, because it is generated in the backend and called via Ajax.

The HTML (report strings) received via Ajax are processed and executed using htmx.process(...).

Additionally, to ensure DRY code, the HTML code required for the modal was added to these four pages as a single-line fragment:
<div th:replace="~{fragments/topicInfoModal :: topicInfoModal}"></div>

@Controller
@RequestMapping("/topics")
public class TopicInfoController {

 @GetMapping("/info/{id}")
    public String getTopicInfo(@PathVariable Long id, Model model) {

    ...

            return "fragments/topicInfoContent :: content";

    }